% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_bins}
\alias{create_bins}
\title{Create bins for the map}
\usage{
create_bins(selected_data, num_bin, selected_year)
}
\arguments{
\item{selected_data}{A reactive object containing the selected_data data.}

\item{num_bin}{Number of bins}
}
\value{
A vector of bins.
}
\description{
Create bins for the map
}
