% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_labels}
\alias{create_labels}
\title{Create labels for the map}
\usage{
create_labels(selected_data, selected_year)
}
\arguments{
\item{selected_data}{A reactive object containing the selected_data data.}
}
\value{
A list of HTML formatted labels.
}
\description{
Create labels for the map
}
